BUG: Changed icon handling in subject hierarchy
Setting icons is not so simple as it should be due to a bug in Qt (https://bugreports.qt-project.org/browse/QTBUG-20248), which raises the need to check if the icon to be set is the same as the current icon of an item. This introduces one more thing to remember to the developer. Thus, instead of having the plugins set icons on items, they just return it (or an empty one), and the model does the check. This way developing plugins is not harder than before.