ENH: Show potential nodes in subject hierarchy tree
Instead of having the potential nodes list on the bottom of the module, those nodes are shown in the subject hierarchy tree on the top-level without icons. They can be shown/hidden with a checkbox under the tree, and can be handled the same way as in the potential list before.