ENH: Log to file
Enabled file logging in CTK error log model. The number of files to keep and the log file paths are stored in the revision settings. The log files are saved in the temp folder with file name Slicer_[revision]_[date]_[time].log. If number of log files is larger than number of files to keep, then the old log file is deleted.