ENH: DICOM tag values in nodes; Calling python plugins from DICOM exporter WIP
DICOM tag values are now added to the nodes on the corresponding level (patient info in patient node etc.). Few extra tags that are potentially to be filled for exporting (mainly from the old export dialog)

Proof of concept that python plugins can be called from the C++ DICOMExportDialog. It will be very hard to get the members of the structures, so C++ classes for DICOMLoadable and DICOMExportable are needed (the loadable one for consistency)