ENH: Get exportable object in export dialog from python plugins WIP
Container objects had to be moved to Subject hierarchy widgets to avoid circular dependency. The only reason a circular dependency would happen is the fact that the export dialog needs to be accessible from subject hierarchy module widget for opening. If this can be done via signal-slot or a simple python call, then this can be avoided and these three classes can be moved to DICOMLib/Widgets where they belong.