STYLE: Rename examine to examineForImport in DICOM plugins
This was done to avoid confusion (there is a new examineForExport method). Backwards compatibility ensured by looking for examine if examineForImport does not exist.
Also change the way the exportables are passed from the export dialog to the DICOM plugins, they are set all at once to accommodate composite exports (when series reference each other)