ENH: Added terminology dialog, selector button, and delegate
The delegate needs to be set to the segment table (and other widgets that will use terminology) to be able to show and change terminology information. The delegate's editor is the new selector button that only shows the recommended color, but when clicked, shows the new selector dialog. The dialog is a simple wrapper dialog class for the navigator widget with select and cancel buttons.